AI commands embark, udisembark and speed now work properly.
Also some bug fixes for embark (was trying to
unset the map twice) and udisembark (was trying
to set the map twice, plus was not cycling
the weapons after the unit left the carrier)

Split the mech_udisembark/mech_sdisembark function(s).
Now mech_udisembark is defined if BT_CARRIERS is
defined, else mech_sdisembark is defined instead
of all those #ifdef/#ifndefs inside of the one
function.



git-svn-id: https://svn.code.sf.net/p/btonline-btech/code/btmux-unstable/trunk@99 e60b412a-5d77-420b-80c2-6cef64185626